Moved erode and dilate to math module
